Unify and improve Excel readout functions
The whole Excel file is read out in one step with the extended function "getPoolDataInDict". "getPoolData", "getZoneArea" and "getZoneVolume" are replaced. All information are stored in one dictionary. 
Number of parameters in the Excel sheet can now be changed without facing issues with the readout.
Furthermore, water zones are not created, when the water area is 0.